--- a/iocore/eventsystem/UnixEThread.cc
+++ b/iocore/eventsystem/UnixEThread.cc
@@ -142,6 +142,97 @@ EThread::process_event(Event *e, int calling_code)
   }
 }
 
+void EThread::process_queue(Que(Event, link) * NegativeQueue)
+{
+  Event *e;
+
+  // Move events from the external thread safe queues to the local queue.
+  EventQueueExternal.dequeue_external();
+
+  // execute all the available external events that have
+  // already been dequeued
+  while ((e = EventQueueExternal.dequeue_local())) {
+    if (e->cancelled) {
+      free_event(e);
+    } else if (!e->timeout_at) { // IMMEDIATE
+      ink_assert(e->period == 0);
+      process_event(e, e->callback_event);
+    } else if (e->timeout_at > 0) { // INTERVAL
+      EventQueue.enqueue(e, cur_time);
+    } else { // NEGATIVE
+      Event *p = nullptr;
+      Event *a = NegativeQueue->head;
+      while (a && a->timeout_at > e->timeout_at) {
+        p = a;
+        a = a->link.next;
+      }
+      if (!a) {
+        NegativeQueue->enqueue(e);
+      } else {
+        NegativeQueue->insert(e, p);
+      }
+    }
+  }
+}
+
+void
+EThread::execute_regular()
+{
+  Event *e;
+  Que(Event, link) NegativeQueue;
+  ink_hrtime next_time = 0;
+
+  // give priority to immediate events
+  for (;;) {
+    if (unlikely(shutdown_event_system == true)) {
+      return;
+    }
+
+    process_queue(&NegativeQueue);
+
+    bool done_one;
+    do {
+      done_one = false;
+      // execute all the eligible internal events
+      EventQueue.check_ready(cur_time, this);
+      while ((e = EventQueue.dequeue_ready(cur_time))) {
+        ink_assert(e);
+        ink_assert(e->timeout_at > 0);
+        if (e->cancelled)
+          free_event(e);
+        else {
+          done_one = true;
+          process_event(e, e->callback_event);
+        }
+      }
+    } while (done_one);
+
+    // execute any negative (poll) events
+    if (NegativeQueue.head) {
+      process_queue(&NegativeQueue);
+
+      // execute poll events
+      while ((e = NegativeQueue.dequeue())) {
+        process_event(e, EVENT_POLL);
+      }
+    }
+
+    next_time             = EventQueue.earliest_timeout();
+    ink_hrtime sleep_time = next_time - Thread::get_hrtime_updated();
+    if (sleep_time > 0) {
+      sleep_time = std::min(sleep_time, HRTIME_MSECONDS(THREAD_MAX_HEARTBEAT_MSECONDS));
+    } else {
+      sleep_time = 0;
+    }
+
+    if (n_ethreads_to_be_signalled) {
+      flush_signals(this);
+    }
+
+    tail_cb->waitForActivity(sleep_time);
+  }
+}
+
 //
 // void  EThread::execute()
 //
@@ -168,128 +259,12 @@ EThread::execute()
 
   switch (tt) {
   case REGULAR: {
-    Event *e;
-    Que(Event, link) NegativeQueue;
-    ink_hrtime next_time = 0;
-
-    // give priority to immediate events
-    for (;;) {
-      if (unlikely(shutdown_event_system == true)) {
-        return;
-      }
-      // execute all the available external events that have
-      // already been dequeued
-      cur_time = Thread::get_hrtime_updated();
-      while ((e = EventQueueExternal.dequeue_local())) {
-        if (e->cancelled) {
-          free_event(e);
-        } else if (!e->timeout_at) { // IMMEDIATE
-          ink_assert(e->period == 0);
-          process_event(e, e->callback_event);
-        } else if (e->timeout_at > 0) { // INTERVAL
-          EventQueue.enqueue(e, cur_time);
-        } else { // NEGATIVE
-          Event *p = nullptr;
-          Event *a = NegativeQueue.head;
-          while (a && a->timeout_at > e->timeout_at) {
-            p = a;
-            a = a->link.next;
-          }
-          if (!a) {
-            NegativeQueue.enqueue(e);
-          } else {
-            NegativeQueue.insert(e, p);
-          }
-        }
-      }
-      bool done_one;
-      do {
-        done_one = false;
-        // execute all the eligible internal events
-        EventQueue.check_ready(cur_time, this);
-        while ((e = EventQueue.dequeue_ready(cur_time))) {
-          ink_assert(e);
-          ink_assert(e->timeout_at > 0);
-          if (e->cancelled) {
-            free_event(e);
-          } else {
-            done_one = true;
-            process_event(e, e->callback_event);
-          }
-        }
-      } while (done_one);
-      // execute any negative (poll) events
-      if (NegativeQueue.head) {
-        if (n_ethreads_to_be_signalled) {
-          flush_signals(this);
-        }
-        // dequeue all the external events and put them in a local
-        // queue. If there are no external events available, don't
-        // do a cond_timedwait.
-        if (!INK_ATOMICLIST_EMPTY(EventQueueExternal.al)) {
-          EventQueueExternal.dequeue_timed(cur_time, next_time, false);
-        }
-        while ((e = EventQueueExternal.dequeue_local())) {
-          if (!e->timeout_at) {
-            process_event(e, e->callback_event);
-          } else {
-            if (e->cancelled) {
-              free_event(e);
-            } else {
-              // If its a negative event, it must be a result of
-              // a negative event, which has been turned into a
-              // timed-event (because of a missed lock), executed
-              // before the poll. So, it must
-              // be executed in this round (because you can't have
-              // more than one poll between two executions of a
-              // negative event)
-              if (e->timeout_at < 0) {
-                Event *p = nullptr;
-                Event *a = NegativeQueue.head;
-                while (a && a->timeout_at > e->timeout_at) {
-                  p = a;
-                  a = a->link.next;
-                }
-                if (!a) {
-                  NegativeQueue.enqueue(e);
-                } else {
-                  NegativeQueue.insert(e, p);
-                }
-              } else {
-                EventQueue.enqueue(e, cur_time);
-              }
-            }
-          }
-        }
-        // execute poll events
-        while ((e = NegativeQueue.dequeue())) {
-          process_event(e, EVENT_POLL);
-        }
-        if (!INK_ATOMICLIST_EMPTY(EventQueueExternal.al)) {
-          EventQueueExternal.dequeue_timed(cur_time, next_time, false);
-        }
-      } else { // Means there are no negative events
-        next_time             = EventQueue.earliest_timeout();
-        ink_hrtime sleep_time = next_time - cur_time;
-
-        if (sleep_time > THREAD_MAX_HEARTBEAT_MSECONDS * HRTIME_MSECOND) {
-          next_time = cur_time + THREAD_MAX_HEARTBEAT_MSECONDS * HRTIME_MSECOND;
-        }
-        // dequeue all the external events and put them in a local
-        // queue. If there are no external events available, do a
-        // cond_timedwait.
-        if (n_ethreads_to_be_signalled) {
-          flush_signals(this);
-        }
-        EventQueueExternal.dequeue_timed(cur_time, next_time, true);
-      }
-    }
+    this->execute_regular();
+    break;
   }
-
   case DEDICATED: {
     break;
   }
-
   default:
     ink_assert(!"bad case value (execute)");
     break;