[SWOC says] [Presentation on these changes.](https://www.dropbox.com/s/udzrj6o9tb8bsxp/Event-Processing-Spring-2016.pdf?dl=0)