@ushachar 2017-09-19
Could you add some details on what this change accomplishes / fixes?