@SolidWallOfCode 2017-09-19
It may be needful to split this in to two PRs - one for the event loop changes, and another for the metric changes.