@mingzym 2017-09-20
I'd like @oknet to review the codes and I am curious of why this modify is taken from the root cause, and how the performance metrics look like before and after this change if you have any.

thanks