@oknet 2017-09-20
For the `EThread::execute()`,

- It is acceptable if an event in the EventQ is delay to callback because It may take a long time to process them one by one.
- The IMM Event should be called back ASAP.
- An event in the NegativeQ would be processed for a long time because It may be blocked at somewhere.
- In order to process imm Event ASAP, the ProtectQ should be checked before EventQ and/or NegativeQ is checked.

But the new `EThread::execute_regular()` did not meet these requirements.