@SolidWallOfCode 2017-09-20
@ushachar Yes, sorry about that. This was originally proposed back at the [Spring 2016 Summit](https://cwiki.apache.org/confluence/display/TS/Presentations+-+2016) - look at the "Event Loop Changes" presentation of mine

The purpose of this is several fold, but primarily to improve performance in low latency applications at variable transaction rates. This has been in production at Yahoo! since 2015 so the basic change are well tested. What this does is remove the condition variable as the stall point for the event loop and just uses the I/O wait for that. As a consequence it also removes the "negative event" used to drive I/O. This makes the event loop much more responsive when there is an I/O stall.