@SolidWallOfCode 2017-09-20
@mingzym - @duke8253 has been doing some performance measurements on this, hopefully he'll post his results. In essence we see improvement in some cases and no change in others, yielding an overall clear positive benefit. I originally did this work to deal with latencies in a low latency application inside Yahoo! where the average transaction time was ~6ms and the 99% value had to be under 20ms. This was not achievable with the existing event loop because the condition variable based wait simply couldn't get active again in that time frame. We observed regular spikes of 40ms latency because of that.  After working with the code, I frankly saw no reason at all for using the condition variable instead of just stalling on the I/O wait. I suspect the condition var was put it before the wake up pipe was added to the `epoll`.