@SolidWallOfCode 2017-09-20
@oknet - the event loop logic you note should be identical to what it was before the PR, the only change in that respect is to move the code into a method, out of the `switch`. Can you point out the exact code changes that you think break the previous model?