@bryancall 2017-09-20
Please squash commits and don't reference internal bug ids in open source.