@oknet 2017-09-21
@SolidWallOfCode 

In my opinion, 

- The Immediately Event is the highest priority and it should be processed ASAP.
- The In, At and Every Event is standard priority and it may not be processed on time
   - These Events are stored in EventQueue.
   - ATS may take a long time to process them one by one.
- The Negative Event is the lowest priority because it could blocks the EThread loop for awhile.
   - Generally, the Negative Event is used to trigger a sub-system.
   - if there is a block operation in a sub-system, It must register a `cb_func` to EventSystem in order to break the block operation and return to the main-system (EventSystem).
   - A sub-system loop may run or not in a loop of main-system.
   - For example: `schedule_imm_signal()`, it create a Immediately Event then push the event into a EThread and break any block operation in that EThread. The Immediately Event is highest priority.

The PR removes the logic of checking PQ between EventQueue and NegativeQ, which may delay the Immediately Event.

```
 +    // execute any negative (poll) events
 +    if (NegativeQueue.head) {
 +      // execute poll events
 +      while ((e = NegativeQueue.dequeue())) {
 +        process_event(e, EVENT_POLL);
 +      }
 +    }
```

The `SCOPED_MUTEX_LOCK` may block the main-system loop for awhile. The main loop should not block for a mutex of sub-system.

```
 +int
 +NetHandler::waitForActivity(ink_hrtime timeout)
 +{
    EventIO *epd = nullptr;
  		  
    NET_INCREMENT_DYN_STAT(net_handler_run_stat);
 +  SCOPED_MUTEX_LOCK(lock, mutex, this->thread);
```

According to your comments:

```
In essence we see improvement in some cases and no change in others, yielding an overall clear positive benefit. I originally did this work to deal with latencies in a low latency application inside Yahoo! where the average transaction time was ~6ms and the 99% value had to be under 20ms.
```

In my opinion, The PR force `NetHandler` runs in every main-system loops, It is the key change to improve the performance.

Can we get the same performance improvement with creating new `process_negative_event()` that use `SCOPED_MUTEX_LOCK` to replace `MUTEX_TRY_LOCK_FOR` in `process_event()` ?

```
void
EThread::process_negative_event(Event *e, int calling_code)
{
  ink_assert((!e->in_the_prot_queue && !e->in_the_priority_queue));
  {
    SCOPED_MUTEX_LOCK(lock, e->mutex, this);
    if (e->cancelled) {
      free_event(e);
      return;
    }
    Continuation *c_temp = e->continuation;
    e->continuation->handleEvent(calling_code, e); 
    ink_assert(!e->in_the_priority_queue);
    ink_assert(c_temp == e->continuation);
  }
  if (!e->in_the_prot_queue && !e->in_the_priority_queue) {
    e->timeout_at = e->period;
    EventQueueExternal.enqueue_local(e);
  }
}
```

```
        // execute poll events
        while ((e = NegativeQueue.dequeue())) {
          process_negative_event(e, EVENT_POLL);
        } 
```