@SolidWallOfCode 2017-09-21
@oknet quick reply on one point - IMHO the key change for better performance is the removal of the condition variable. Experiments leading to this change indicated that was where the event loop would get "stuck" and not respond promptly to I/O events because it would wait on external signals / events, not I/O readiness.

I need to ponder your other points.