@oknet 2017-09-23
@SolidWallOfCode Do you means `might_have_data` ?