@SolidWallOfCode 2017-09-26
[~oknet]

By the condition variable, I do mean that in `ProtectedQueue::dequeue_wait`.

I think your concerns about UDP traffic are misplaced. Looking at the code, there is already a
different set of threads for UDP so changes to how the TCP based `NetHandler` instances are handled
will have no effect. UDP can continue to use negative events just as before. Note that if no
explicit tail handler is set, the default is the same logic as before the changes. See
`DEFAULT_TAIL_HANDLER` which when it waits, calls `ProtectedQueue::wait` which is the wait part of
`ProtectedQueue::dequeue_timed`. I think the UDP handling should be converted to use the tail wait
as the TCP style does, but it's not critical.

Having spoken with [~duke8253] we decided to put the secondary external queue dequeue back in, past
the main event dispatch loop. From my point of view I don't think that has any benefits but it seems
low cost so why not? Fei (duke8253) will extract that code into a method and put calls to it the
equivalent places it is in the current code.