@oknet 2017-09-26
@SolidWallOfCode 

I can not found `ProtectedQueue::dequeue_wait`. Do you means `ProtectedQueue::dequeue_timed` ?