@duke8253 2017-09-29
@oknet new test results

master branch
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.38ms    1.31ms  63.55ms   72.41%
    Req/Sec     2.63k   343.43     5.45k    71.37%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.17ms
 75.000%    3.08ms
 90.000%    4.10ms
 99.000%    5.88ms
 99.900%   11.18ms
 99.990%   15.85ms
 99.999%   40.13ms
100.000%   63.58ms
#[Mean    =        2.384, StdDeviation   =        1.313]
#[Max     =       63.552, Total count    =      5787453]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994221 requests in 5.00m, 6.61GB read
Requests/sec:  19980.78
Transfer/sec:     22.57MB

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.41ms    1.38ms  63.62ms   73.82%
    Req/Sec     2.63k   347.56     7.67k    70.53%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.18ms
 75.000%    3.10ms
 90.000%    4.18ms
 99.000%    6.05ms
 99.900%   11.98ms
 99.990%   23.69ms
 99.999%   40.06ms
100.000%   63.65ms
#[Mean    =        2.412, StdDeviation   =        1.376]
#[Max     =       63.616, Total count    =      5787450]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994225 requests in 5.00m, 6.61GB read
Requests/sec:  19980.81
Transfer/sec:     22.55MB

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.34ms    1.24ms  34.21ms   72.36%
    Req/Sec     2.63k   329.08     4.55k    69.89%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.15ms
 75.000%    3.02ms
 90.000%    3.93ms
 99.000%    5.78ms
 99.900%   11.17ms
 99.990%   14.37ms
 99.999%   22.37ms
100.000%   34.24ms
#[Mean    =        2.344, StdDeviation   =        1.241]
#[Max     =       34.208, Total count    =      5787439]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994216 requests in 5.00m, 6.61GB read
Requests/sec:  19980.72
Transfer/sec:     22.56MB
```
event loop changes:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.40ms    1.30ms  39.81ms   71.58%
    Req/Sec     2.63k   339.81     4.78k    72.02%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.18ms
 75.000%    3.11ms
 90.000%    4.13ms
 99.000%    5.90ms
 99.900%   11.02ms
 99.990%   14.82ms
 99.999%   24.21ms
100.000%   39.84ms
#[Mean    =        2.396, StdDeviation   =        1.305]
#[Max     =       39.808, Total count    =      5787464]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994222 requests in 5.00m, 6.62GB read
Requests/sec:  19980.72
Transfer/sec:     22.58MB

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.42ms    1.30ms  54.05ms   71.95%
    Req/Sec     2.63k   340.90     5.00k    71.26%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.21ms
 75.000%    3.12ms
 90.000%    4.11ms
 99.000%    5.92ms
 99.900%   11.25ms
 99.990%   15.45ms
 99.999%   31.41ms
100.000%   54.08ms
#[Mean    =        2.419, StdDeviation   =        1.298]
#[Max     =       54.048, Total count    =      5787451]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994227 requests in 5.00m, 6.62GB read
Requests/sec:  19980.85
Transfer/sec:     22.58MB

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.35ms    1.24ms  26.50ms   72.23%
    Req/Sec     2.63k   329.58     4.67k    69.81%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.15ms
 75.000%    3.03ms
 90.000%    3.94ms
 99.000%    5.81ms
 99.900%   10.89ms
 99.990%   14.34ms
 99.999%   22.11ms
100.000%   26.51ms
#[Mean    =        2.350, StdDeviation   =        1.242]
#[Max     =       26.496, Total count    =      5787433]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994220 requests in 5.00m, 6.62GB read
Requests/sec:  19980.70
Transfer/sec:     22.58MB
```

PR #2572:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.36ms    1.42ms  66.50ms   72.98%
    Req/Sec     2.50k     2.43     2.51k    86.95%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.09ms
 75.000%    3.12ms
 90.000%    4.18ms
 99.000%    6.57ms
 99.900%   10.73ms
 99.990%   17.57ms
 99.999%   42.59ms
100.000%   66.56ms
#[Mean    =        2.357, StdDeviation   =        1.417]
#[Max     =       66.496, Total count    =      5787419]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994171 requests in 5.00m, 6.61GB read
Requests/sec:  19980.51
Transfer/sec:     22.55MB

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.57ms    1.43ms  47.42ms   71.93%
    Req/Sec     2.63k   318.92     5.00k    70.40%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.36ms
 75.000%    3.39ms
 90.000%    4.35ms
 99.000%    6.24ms
 99.900%   12.65ms
 99.990%   17.47ms
 99.999%   27.22ms
100.000%   47.46ms
#[Mean    =        2.568, StdDeviation   =        1.428]
#[Max     =       47.424, Total count    =      5787451]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994229 requests in 5.00m, 6.61GB read
Requests/sec:  19980.72
Transfer/sec:     22.56MB

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.57ms    1.43ms  58.30ms   72.09%
    Req/Sec     2.63k   316.62     5.70k    70.82%
  Latency Distribution (HdrHistogram - Recorded Latency)
 50.000%    2.36ms
 75.000%    3.40ms
 90.000%    4.36ms
 99.000%    6.23ms
 99.900%   12.69ms
 99.990%   19.34ms
 99.999%   33.60ms
100.000%   58.33ms
#[Mean    =        2.574, StdDeviation   =        1.435]
#[Max     =       58.304, Total count    =      5787447]
#[Buckets =           27, SubBuckets     =         2048]
----------------------------------------------------------
  5994229 requests in 5.00m, 6.62GB read
Requests/sec:  19980.76
Transfer/sec:     22.58MB
```